refactor(devices): replace mobile-sync dialogs with inline panels
Remove the DeviceSettingsDialog, MobileSyncCredentialModal, and
MobileSyncDeviceDialog modal stack in favor of panel-based components
(LocalDevicePanel, MobileDevicePanel, PeerDetailPanel) rendered inline on
DevicesPage. One-time credentials are now echoed in MobileDevicePanel's
fresh state instead of a dedicated modal. Adds shared UI primitives
(StatusDot, CopyIconButton, PanelFactRow, MobileSyncInstallHelper) and
updates i18n strings accordingly.
