+++ src/pages/DevicesPage.tsx
@@ -1,37 +1,28 @@
 /**
- * DevicesPage —— "Hero + Tab + 卡片网格" 布局（参考 DevicesPagePrototypeV2）。
+ * DevicesPage: master-detail layout ("Ledger" treatment).
  *
- * 三层结构：
+ *   ┌─ list column ──────┬─ detail pane ──────────────────────────┐
+ *   │ 设备        2/3 在线 │  Windows 工作站            [取消配对]  │
+ *   │ ── 本机 ──────────  │  ● 在线 · 局域网直连                    │
+ *   │ ● MacBook Pro      │  ─────────────────────────────────────  │
+ *   │ ── 已配对设备 ────  │  PEER ID   通道   地址                  │
+ *   │ ● Windows 工作站 ◀ │  ─────────────────────────────────────  │
+ *   │ ● Arch VM          │  同步设置（开关 + 内容类型）             │
+ *   │ ── 移动同步 ──  ⚙  │                                         │
+ *   │ ● iPhone 15 Pro    │                                         │
+ *   │ ＋ 添加设备         │                                         │
+ *   └────────────────────┴─────────────────────────────────────────┘
  *
- *   ┌─ Hero ──────────────────────────────────────────────────────┐
- *   │ 本设备图标 + 名字 + peerId + 切换空间 chip │ 配对 / 同步 pill │
- *   ├─ Tabs (P2P / Mobile) ───────────────────────────────────────┤
- *   │  ┌──────┐ ┌──────┐ ┌──────┐                                  │
- *   │  │ peer │ │ peer │ │  +   │                                  │
- *   │  └──────┘ └──────┘ └──────┘                                  │
- *   └─────────────────────────────────────────────────────────────┘
- *
- * 所有 dialog / modal 复用现成实现（AddDeviceDialog / DeviceSettingsDialog /
- * MobileSync* 系列），本页只重写视觉骨架与卡片排版。`refreshPresence`
- * 现在只在挂载和页面从隐藏切回可见时触发一次，常态依赖 daemon 主动推送
- * 的 `peers.changed` ws 事件驱动 UI（详见下方 useEffect 的注释）。
+ * Selecting a device renders its detail inline (LocalDevicePanel /
+ * PeerDetailPanel / MobileDevicePanel) instead of stacking dialogs.
+ * Flow dialogs (add / enable / credential echo / unpair / revoke /
+ * mobile-sync settings) remain modal. `refreshPresence` fires only on
+ * mount and on visibility regain; steady-state updates ride the
+ * daemon-pushed `peers.changed` ws events.
  */
 
-import {
-  ArrowRightLeft,
-  Cable,
-  CheckCircle2,
-  ChevronRight,
-  Pause,
-  Plus,
-  RefreshCw,
-  Settings2,
-  ShieldCheck,
-  Smartphone,
-  Wifi,
-  WifiOff,
-} from 'lucide-react'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import { Plus, Settings2 } from 'lucide-react'
+import React, { useCallback, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { refreshPresence } from '@/api/daemon'
 import type { SpaceMember } from '@/api/daemon/members'
@@ -49,13 +40,12 @@ import {
 import AddDeviceDialog from '@/components/device/AddDeviceDialog'
 import AddMobileSyncDeviceDialog from '@/components/device/AddMobileSyncDeviceDialog'
 import { deriveBadgeKind } from '@/components/device/connection-channel-utils'
-import { getDeviceIcon } from '@/components/device/device-utils'
-import DeviceSettingsDialog from '@/components/device/DeviceSettingsDialog'
 import EnableMobileSyncDialog from '@/components/device/EnableMobileSyncDialog'
-import MobileSyncCredentialModal from '@/components/device/MobileSyncCredentialModal'
-import MobileSyncDeviceDialog from '@/components/device/MobileSyncDeviceDialog'
+import LocalDevicePanel from '@/components/device/LocalDevicePanel'
+import MobileDevicePanel from '@/components/device/MobileDevicePanel'
 import MobileSyncSettingsDialog from '@/components/device/MobileSyncSettingsDialog'
-import SwitchSpaceDialog from '@/components/device/SwitchSpaceDialog'
+import PeerDetailPanel from '@/components/device/PeerDetailPanel'
+import StatusDot, { type StatusDotTone } from '@/components/device/StatusDot'
 import UnpairAlertDialog from '@/components/device/UnpairAlertDialog'
 import { Alert, AlertDescription } from '@/components/ui/alert'
 import {
@@ -69,15 +59,19 @@ import {
   AlertDialogTitle,
 } from '@/components/ui/alert-dialog'
 import { Button } from '@/components/ui/button'
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu'
 import { ScrollArea } from '@/components/ui/scroll-area'
 import { Skeleton } from '@/components/ui/skeleton'
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { toast } from '@/components/ui/toast'
 import type { PeerSnapshotPayloadItem, PeersChangedPayload } from '@/hooks/useDaemonEvents'
 import { useSetting } from '@/hooks/useSetting'
 import { daemonWs } from '@/lib/daemon-ws'
 import { createLogger } from '@/lib/logger'
-import { formatPeerIdForDisplay } from '@/lib/utils'
 import { cn } from '@/lib/utils'
 import { useAppDispatch, useAppSelector } from '@/store/hooks'
 import {
@@ -90,28 +84,46 @@ import {
 
 const log = createLogger('devices-page')
 
-// ────────────────────────────────────────────────────────────────
-// 顶层页面
-// ────────────────────────────────────────────────────────────────
+type Selection = { kind: 'local' } | { kind: 'peer'; id: string } | { kind: 'mobile'; id: string }
+
+/** A mobile device counts as "recently active" within this window. */
+const MOBILE_ACTIVE_WINDOW_MS = 10 * 60 * 1000
 
 const DevicesPage: React.FC = () => {
+  const { t } = useTranslation()
   const dispatch = useAppDispatch()
+
+  const {
+    localDevice,
+    localDeviceLoading,
+    localDeviceError,
+    spaceMembers: rawSpaceMembers,
+    spaceMembersError,
+  } = useAppSelector(state => state.devices)
+
+  const peers = localDevice
+    ? rawSpaceMembers.filter(d => d.peerId !== localDevice.peerId)
+    : rawSpaceMembers
+  const onlineCount = peers.filter(p => p.connected).length
+
+  const { setting } = useSetting()
+  const syncActive = setting?.sync.autoSync !== false
+  const globalAutoSyncOff = setting?.sync.autoSync === false
+  const globalFileSyncOff = setting?.fileSync?.fileSyncEnabled === false
+  const lanOnlyActive = setting?.network?.allowRelayFallback === false
 
   useEffect(() => {
     dispatch(fetchLocalDeviceInfo())
     dispatch(fetchSpaceMembers())
 
-    // 上线感知由 daemon 的 PeerKeepAliveWorker 推送驱动：inbound presence
-    // Online → outbound dial → peers.changed ws → 前端切亮（~1s）。
-    // 离线感知交给 daemon 自身的 25s keepalive tick + QUIC idle watchdog。
-    // 前端只在两个时刻主动拉一次 presence:
-    //   1. 页面首次挂载 —— warm 一下 UI，避免 ws 推送之前显示陈旧状态。
-    //   2. 标签页从隐藏切回可见 —— 用户回到前台时给一个兜底快照。
-    // 不再 setInterval polling，关闭 Devices 页时 daemon 完全静默。
+    // Presence awareness is push-driven by the daemon's PeerKeepAliveWorker
+    // (inbound presence Online → outbound dial → peers.changed ws). The
+    // frontend only pulls presence twice: on first mount (warm the UI) and
+    // when the tab becomes visible again (safety snapshot). No polling.
     const probe = () => {
       refreshPresence().catch(err => {
-        // setup 未完成 / daemon 未就绪时 refresh_presence 会 5xx；
-        // 不影响后续推送链路，warn 即可。
+        // refresh_presence 5xxs while setup is incomplete / daemon not
+        // ready; the push pipeline is unaffected, so warn is enough.
         log.warn({ err }, 'presence refresh failed')
       })
     }
@@ -128,247 +140,6 @@ const DevicesPage: React.FC = () => {
     }
   }, [dispatch])
 
-  return (
-    <div className="flex flex-col h-full relative">
-      <div className="flex-1 overflow-hidden relative">
-        <ScrollArea className="h-full">
-          <div className="space-y-8 px-6 pb-12 pt-6">
-            <HeroSection />
-            <DeviceTabs />
-          </div>
-        </ScrollArea>
-      </div>
-    </div>
-  )
-}
-
-export default DevicesPage
-
-// ────────────────────────────────────────────────────────────────
-// Hero 区
-// ────────────────────────────────────────────────────────────────
-
-const HeroSection: React.FC = () => {
-  const { t } = useTranslation()
-  const dispatch = useAppDispatch()
-  const { localDevice, localDeviceLoading, localDeviceError, spaceMembers } = useAppSelector(
-    state => state.devices
-  )
-  const { setting } = useSetting()
-  const syncActive = setting?.sync.autoSync !== false
-
-  const peers = localDevice
-    ? spaceMembers.filter(d => d.peerId !== localDevice.peerId)
-    : spaceMembers
-  const pairedCount = peers.length
-  const onlineCount = peers.filter(p => p.connected).length
-
-  if (localDeviceError) {
-    return (
-      <Alert variant="destructive">
-        <AlertDescription className="flex items-center gap-3">
-          <span className="flex-1">{localDeviceError}</span>
-          <Button
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => {
-              dispatch(clearLocalDeviceError())
-              dispatch(fetchLocalDeviceInfo())
-            }}
-            title={t('devices.list.actions.retry')}
-          >
-            <RefreshCw className="h-4 w-4" />
-          </Button>
-        </AlertDescription>
-      </Alert>
-    )
-  }
-
-  if (localDeviceLoading && localDevice === null) {
-    return (
-      <div className="grid grid-cols-1 gap-3 lg:grid-cols-[1fr_auto]">
-        <div className="rounded-2xl border border-border/60 bg-card px-6 py-6">
-          <div className="flex items-center gap-5">
-            <Skeleton className="h-16 w-16 rounded-2xl" />
-            <div className="flex flex-1 flex-col gap-2">
-              <Skeleton className="h-3 w-12" />
-              <Skeleton className="h-5 w-40" />
-              <Skeleton className="h-3 w-32" />
-            </div>
-          </div>
-        </div>
-        <div className="grid grid-cols-2 gap-3 lg:flex lg:w-[220px] lg:flex-col">
-          <Skeleton className="h-[68px] rounded-xl" />
-          <Skeleton className="h-[68px] rounded-xl" />
-        </div>
-      </div>
-    )
-  }
-
-  if (!localDevice) return null
-
-  const Icon = getDeviceIcon(localDevice.deviceName)
-
-  return (
-    <div className="grid grid-cols-1 gap-3 lg:grid-cols-[1fr_auto]">
-      {/* ── 本机 hero (左) ────────────────────────────────────── */}
-      <div className="relative overflow-hidden rounded-2xl border border-border/60 bg-card px-6 py-6">
-        <div
-          aria-hidden
-          className="pointer-events-none absolute inset-0 opacity-[0.04]"
-          style={{
-            backgroundImage: 'radial-gradient(circle at 1px 1px, currentColor 1px, transparent 0)',
-            backgroundSize: '16px 16px',
-          }}
-        />
-        <div className="relative flex items-center gap-5">
-          <div className="flex h-16 w-16 shrink-0 items-center justify-center rounded-2xl bg-success/15 text-success shadow-sm ring-1 ring-success/20">
-            {/* eslint-disable-next-line react-hooks/static-components -- `getDeviceIcon` returns a stable lucide icon reference keyed on deviceName, not a freshly-created component */}
-            <Icon className="h-8 w-8" />
-          </div>
-          <div className="min-w-0 flex-1">
-            <div className="flex items-center gap-2">
-              <span className="text-[11px] uppercase tracking-wider text-muted-foreground">
-                {t('devices.thisDevice.title')}
-              </span>
-              <span className="inline-flex h-1.5 w-1.5 rounded-full bg-success" />
-            </div>
-            <h2 className="mt-1 truncate text-xl font-semibold leading-tight text-foreground">
-              {localDevice.deviceName}
-            </h2>
-            <p className="mt-1 truncate font-mono text-xs text-muted-foreground/80">
-              {formatPeerIdForDisplay(localDevice.peerId)}
-            </p>
-          </div>
-
-          <div className="flex shrink-0 items-start gap-2 self-start pt-1">
-            <SpaceChip />
-          </div>
-        </div>
-      </div>
-
-      {/* ── 右侧 stat 列 ───────────────────────────────────────── */}
-      <div className="grid grid-cols-2 gap-3 lg:flex lg:w-[220px] lg:flex-col">
-        <StatPill
-          icon={Wifi}
-          label={t('devices.pairedDevices.title')}
-          value={`${onlineCount}/${pairedCount}`}
-          sublabel={
-            pairedCount === 0
-              ? t('devices.pairedDevices.subtitleEmpty')
-              : t('devices.thisDevice.onlineCount', { count: onlineCount })
-          }
-          accent={onlineCount > 0 ? 'emerald' : 'muted'}
-        />
-        <StatPill
-          icon={syncActive ? CheckCircle2 : Pause}
-          label={t('devices.thisDevice.syncStatusLabel', { defaultValue: '同步状态' })}
-          value={
-            syncActive ? t('devices.thisDevice.syncActive') : t('devices.thisDevice.syncPaused')
-          }
-          sublabel={
-            syncActive
-              ? t('devices.thisDevice.syncRunningHint', { defaultValue: '正在自动同步' })
-              : t('devices.syncPaused.goToSettings')
-          }
-          accent={syncActive ? 'emerald' : 'amber'}
-        />
-      </div>
-    </div>
-  )
-}
-
-const SpaceChip: React.FC = () => {
-  const { t } = useTranslation()
-  const [open, setOpen] = useState(false)
-  return (
-    <>
-      <button
-        type="button"
-        onClick={() => setOpen(true)}
-        className="inline-flex items-center gap-1.5 rounded-full px-2.5 py-0.5 text-[11px] font-medium text-muted-foreground ring-1 ring-border/60 transition-colors hover:bg-background hover:text-foreground hover:ring-border"
-      >
-        <ArrowRightLeft className="h-3 w-3" />
-        <span>{t('devices.switchSpace.button')}</span>
-      </button>
-      <SwitchSpaceDialog open={open} onOpenChange={setOpen} />
-    </>
-  )
-}
-
-type StatAccent = 'emerald' | 'amber' | 'muted'
-
-interface StatPillProps {
-  icon: typeof Wifi
-  label: string
-  value: string
-  sublabel: string
-  accent: StatAccent
-}
-
-const STAT_ACCENT: Record<StatAccent, { icon: string; value: string }> = {
-  emerald: {
-    icon: 'bg-success/10 text-success',
-    value: 'text-success',
-  },
-  amber: {
-    icon: 'bg-warning/10 text-warning',
-    value: 'text-warning',
-  },
-  muted: {
-    icon: 'bg-muted text-muted-foreground',
-    value: 'text-foreground',
-  },
-}
-
-const StatPill: React.FC<StatPillProps> = ({ icon: Icon, label, value, sublabel, accent }) => {
-  const a = STAT_ACCENT[accent]
-  return (
-    <div className="flex items-center gap-3 rounded-xl border border-border/60 bg-card px-4 py-3">
-      <div className={cn('flex h-9 w-9 shrink-0 items-center justify-center rounded-lg', a.icon)}>
-        <Icon className="h-4 w-4" />
-      </div>
-      <div className="min-w-0 flex-1">
-        <p className="text-[10px] uppercase tracking-wider text-muted-foreground">{label}</p>
-        <p className={cn('truncate text-sm font-semibold', a.value)}>{value}</p>
-        <p className="truncate text-[11px] text-muted-foreground/80">{sublabel}</p>
-      </div>
-    </div>
-  )
-}
-
-// ────────────────────────────────────────────────────────────────
-// 设备 tabs（P2P / Mobile）
-// ────────────────────────────────────────────────────────────────
-
-type TabKey = 'p2p' | 'mobile'
-
-const DeviceTabs: React.FC = () => {
-  const { t } = useTranslation()
-  const [tab, setTab] = useState<TabKey>('p2p')
-
-  // ── P2P 状态 ─────────────────────────────────────────────────
-  const dispatch = useAppDispatch()
-  const {
-    spaceMembers: rawSpaceMembers,
-    spaceMembersError,
-    localDevice,
-  } = useAppSelector(state => state.devices)
-  const peers = localDevice
-    ? rawSpaceMembers.filter(d => d.peerId !== localDevice.peerId)
-    : rawSpaceMembers
-
-  const [addP2PDialogOpen, setAddP2PDialogOpen] = useState(false)
-  const [selectedPeerId, setSelectedPeerId] = useState<string | null>(null)
-  const [peerDialogOpen, setPeerDialogOpen] = useState(false)
-  const [unpairDialogOpen, setUnpairDialogOpen] = useState(false)
-  const [unpairTargetId, setUnpairTargetId] = useState<string | null>(null)
-
-  const { setting } = useSetting()
-  const globalAutoSyncOff = setting?.sync.autoSync === false
-  const globalFileSyncOff = setting?.fileSync?.fileSyncEnabled === false
-  const lanOnlyActive = setting?.network?.allowRelayFallback === false
-
   useEffect(() => {
     const handler = (event: { topic: string; eventType: string; payload: unknown }) => {
       if (event.topic !== 'peers') return
@@ -384,10 +155,50 @@ const DeviceTabs: React.FC = () => {
     return unsub
   }, [dispatch])
 
-  const handleSelectPeer = useCallback((peerId: string) => {
-    setSelectedPeerId(peerId)
-    setPeerDialogOpen(true)
-  }, [])
+  // ── mobile devices state ──────────────────────────────────────
+  const {
+    devices: mobileDevices,
+    devicesError: mobileDevicesError,
+    settings: mobileSettings,
+    addDialogOpen,
+    settingsSheetOpen,
+    enableConfirmOpen,
+    revokeTarget,
+    revokeBusy,
+    actions: mobileActions,
+  } = useMobileDevices()
+
+  // ── selection ────────────────────────────────────────────────
+  const [selection, setSelection] = useState<Selection>({ kind: 'local' })
+  // One-time credentials from a just-completed add. Handed to the selected
+  // MobileDevicePanel so it shows the pairing QR + credentials + install helper
+  // inline (the old post-registration modal, retired). Cleared as soon as the
+  // user navigates to any other device so the one-time secret can't linger.
+  const [pendingCredential, setPendingCredential] = useState<RegisterMobileDeviceResult | null>(
+    null
+  )
+  useEffect(() => {
+    if (
+      pendingCredential &&
+      !(selection.kind === 'mobile' && selection.id === pendingCredential.deviceId)
+    ) {
+      setPendingCredential(null)
+    }
+  }, [selection, pendingCredential])
+  const selectedPeer =
+    selection.kind === 'peer' ? peers.find(p => p.peerId === selection.id) : undefined
+  const selectedMobile =
+    selection.kind === 'mobile' ? mobileDevices.find(d => d.deviceId === selection.id) : undefined
+  // A vanished selection (unpaired / revoked device) falls back to local.
+  const effectiveSelection: Selection =
+    (selection.kind === 'peer' && !selectedPeer) || (selection.kind === 'mobile' && !selectedMobile)
+      ? { kind: 'local' }
+      : selection
+
+  // ── p2p dialogs ──────────────────────────────────────────────
+  const [addP2PDialogOpen, setAddP2PDialogOpen] = useState(false)
+  const [unpairDialogOpen, setUnpairDialogOpen] = useState(false)
+  const [unpairTargetId, setUnpairTargetId] = useState<string | null>(null)
 
   const handleUnpairRequest = (peerId: string) => {
     setUnpairTargetId(peerId)
@@ -400,163 +211,228 @@ const DeviceTabs: React.FC = () => {
       await unpairDevice(unpairTargetId)
       dispatch(fetchSpaceMembers())
       setUnpairDialogOpen(false)
-      setPeerDialogOpen(false)
       setUnpairTargetId(null)
     } catch (error) {
       log.error({ err: error }, 'failed to unpair device')
     }
   }
 
-  const selectedPeer = peers.find(d => d.peerId === selectedPeerId)
   const unpairTargetDevice = peers.find(d => d.peerId === unpairTargetId)
 
-  // ── Mobile 状态 ──────────────────────────────────────────────
-  const {
-    devices: mobileDevices,
-    devicesError: mobileDevicesError,
-    settings: mobileSettings,
-    addDialogOpen,
-    settingsSheetOpen,
-    enableConfirmOpen,
-    credentialPayload,
-    revokeTarget,
-    revokeBusy,
-    detailDevice: mobileDetailDevice,
-    actions: mobileActions,
-  } = useMobileDevices()
-
-  // ── 渲染 ──────────────────────────────────────────────────────
   return (
-    <>
-      <Tabs value={tab} onValueChange={value => setTab(value as TabKey)} className="w-full">
-        <div className="flex items-center justify-between gap-3 border-b border-border/50">
-          <TabsList variant="line" className="h-11 gap-8 bg-transparent p-0">
-            <TabsTrigger value="p2p" className="gap-2 px-0 text-sm font-medium">
-              <ShieldCheck className="h-4 w-4" />
-              {t('devices.pairedDevices.title')}
-              <CountChip count={peers.length} active={tab === 'p2p'} />
-            </TabsTrigger>
-            <TabsTrigger value="mobile" className="gap-2 px-0 text-sm font-medium">
-              <Smartphone className="h-4 w-4" />
-              {t('devices.mobileSync.title')}
-              <CountChip count={mobileDevices.length} active={tab === 'mobile'} />
-            </TabsTrigger>
-          </TabsList>
-
-          <div className="flex items-center gap-1.5 pb-1">
-            {/* key 上带 tab 名是为了让 tab 切换时 React 卸载旧按钮、挂载新按钮 —— */}
-            {/* 否则两个分支被 reconciliation 复用同一 DOM,Button 的 transition-all */}
-            {/* 会把 bg-primary → 透明的颜色变化做一次过渡,用户看到 primary 闪一下。 */}
-            {tab === 'p2p' ? (
-              <Button
-                key="p2p-add"
-                variant="default"
-                size="sm"
-                onClick={() => setAddP2PDialogOpen(true)}
-              >
-                <Plus className="h-3.5 w-3.5" />
-                {t('devices.list.actions.addDevice')}
-              </Button>
-            ) : (
-              <>
-                <Button
-                  key="mobile-configure"
-                  variant="ghost"
-                  size="sm"
-                  onClick={mobileActions.openSettings}
-                >
-                  <Settings2 className="h-3.5 w-3.5" />
-                  {t('devices.mobileSync.configure')}
-                </Button>
-                <Button
-                  key="mobile-add"
-                  variant="default"
-                  size="sm"
-                  onClick={mobileActions.handleAddClick}
-                  disabled={mobileSettings?.lanListenerError != null}
-                  title={
-                    mobileSettings?.lanListenerError
-                      ? t('devices.mobileSync.statusBar.bindFailed', {
-                          reason: mobileSettings.lanListenerError,
-                        })
-                      : undefined
-                  }
-                >
-                  <Plus className="h-3.5 w-3.5" />
-                  {t('devices.mobileSync.list.addButton')}
-                </Button>
-              </>
-            )}
-          </div>
+    <div className="flex h-full">
+      {/* ── list column ───────────────────────────────────────── */}
+      <aside className="flex w-60 shrink-0 flex-col border-r border-border/50">
+        <div className="px-4 pb-2 pt-5">
+          <h2 className="text-base font-semibold tracking-tight text-foreground">
+            {t('devices.panel.listTitle')}
+          </h2>
+          <p className="mt-0.5 text-[11px] text-muted-foreground">
+            {t('devices.panel.counts', {
+              online: onlineCount,
+              total: peers.length,
+              mobile: mobileDevices.length,
+            })}
+          </p>
         </div>
 
-        <TabsContent value="p2p" className="mt-6">
-          {spaceMembersError ? (
-            <Alert variant="destructive">
-              <AlertDescription className="flex items-center gap-3">
-                <span className="flex-1">{spaceMembersError}</span>
-                <Button
-                  variant="ghost"
-                  size="icon-sm"
-                  onClick={() => {
-                    dispatch(clearSpaceMembersError())
-                    dispatch(fetchSpaceMembers())
-                  }}
+        <ScrollArea className="min-h-0 flex-1">
+          <div className="flex flex-col px-2 pb-3">
+            {(spaceMembersError || mobileDevicesError) && (
+              <Alert variant="destructive" className="mx-1 my-2">
+                <AlertDescription className="flex flex-col gap-2 text-xs">
+                  <span>{spaceMembersError ?? mobileDevicesError}</span>
+                  <Button
+                    variant="outline"
+                    size="sm"
+                    className="self-start"
+                    onClick={() => {
+                      if (spaceMembersError) {
+                        dispatch(clearSpaceMembersError())
+                        dispatch(fetchSpaceMembers())
+                      }
+                      if (mobileDevicesError) {
+                        mobileActions.reload()
+                      }
+                    }}
+                  >
+                    {t('devices.list.actions.retry')}
+                  </Button>
+                </AlertDescription>
+              </Alert>
+            )}
+
+            <SectionLabel label={t('devices.thisDevice.title')} />
+            {localDevice ? (
+              <DeviceListItem
+                name={localDevice.deviceName}
+                tone={syncActive ? 'success' : 'off'}
+                trailing={t('devices.panel.localBadge')}
+                selected={effectiveSelection.kind === 'local'}
+                onSelect={() => setSelection({ kind: 'local' })}
+              />
+            ) : (
+              <div className="px-2.5 py-2">
+                {localDeviceError ? (
+                  <button
+                    type="button"
+                    className="text-left text-xs text-destructive underline underline-offset-2"
+                    onClick={() => {
+                      dispatch(clearLocalDeviceError())
+                      dispatch(fetchLocalDeviceInfo())
+                    }}
+                  >
+                    {t('devices.list.actions.retry')}
+                  </button>
+                ) : (
+                  <Skeleton className="h-5 w-32" />
+                )}
+              </div>
+            )}
+
+            <SectionLabel label={t('devices.pairedDevices.title')} />
+            {peers.map(peer => (
+              <DeviceListItem
+                key={peer.peerId}
+                name={peer.deviceName || t('devices.list.labels.unknownDevice')}
+                tone={peerDotTone(peer, lanOnlyActive)}
+                dimmed={!peer.connected}
+                selected={
+                  effectiveSelection.kind === 'peer' && effectiveSelection.id === peer.peerId
+                }
+                onSelect={() => setSelection({ kind: 'peer', id: peer.peerId })}
+              />
+            ))}
+            {peers.length === 0 && !spaceMembersError && (
+              <p className="px-2.5 py-1.5 text-[11px] text-muted-foreground/70">
+                {t('devices.pairedDevices.subtitleEmpty')}
+              </p>
+            )}
+
+            <SectionLabel
+              label={t('devices.mobileSync.title')}
+              trailing={
+                <button
+                  type="button"
+                  aria-label={t('devices.mobileSync.configure')}
+                  title={t('devices.mobileSync.configure')}
+                  className="rounded-md p-0.5 text-muted-foreground/70 transition-colors hover:text-foreground"
+                  onClick={mobileActions.openSettings}
                 >
-                  <RefreshCw className="h-4 w-4" />
-                </Button>
-              </AlertDescription>
-            </Alert>
-          ) : (
-            <PeerGrid
-              peers={peers}
+                  <Settings2 className="size-3.5" />
+                </button>
+              }
+            />
+            {mobileDevices.map(mobile => (
+              <DeviceListItem
+                key={mobile.deviceId}
+                name={mobile.label}
+                tone={mobileDotTone(mobile)}
+                dimmed={mobileDotTone(mobile) === 'off'}
+                selected={
+                  effectiveSelection.kind === 'mobile' && effectiveSelection.id === mobile.deviceId
+                }
+                onSelect={() => setSelection({ kind: 'mobile', id: mobile.deviceId })}
+              />
+            ))}
+            {mobileDevices.length === 0 && !mobileDevicesError && (
+              <p className="px-2.5 py-1.5 text-[11px] text-muted-foreground/70">
+                {t('devices.mobileSync.list.empty.title')}
+              </p>
+            )}
+          </div>
+        </ScrollArea>
+
+        <div className="border-t border-border/50 p-2">
+          <DropdownMenu>
+            <DropdownMenuTrigger asChild>
+              <Button
+                variant="ghost"
+                size="sm"
+                className="w-full justify-start gap-2 text-muted-foreground hover:text-foreground"
+              >
+                <Plus className="size-3.5" />
+                {t('devices.panel.addMenu.trigger')}
+              </Button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent align="start" className="w-52">
+              <DropdownMenuItem onSelect={() => setAddP2PDialogOpen(true)}>
+                {t('devices.panel.addMenu.p2p')}
+              </DropdownMenuItem>
+              <DropdownMenuItem
+                disabled={mobileSettings?.lanListenerError != null}
+                onSelect={mobileActions.handleAddClick}
+              >
+                {t('devices.panel.addMenu.mobile')}
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        </div>
+      </aside>
+
+      {/* ── detail pane ───────────────────────────────────────── */}
+      <main className="min-w-0 flex-1">
+        <ScrollArea className="h-full">
+          {effectiveSelection.kind === 'local' &&
+            (localDevice ? (
+              <LocalDevicePanel localDevice={localDevice} memberCount={peers.length + 1} />
+            ) : localDeviceError ? (
+              <div className="mx-auto w-full max-w-2xl px-8 py-8">
+                <Alert variant="destructive">
+                  <AlertDescription className="flex items-center gap-3">
+                    <span className="flex-1">{localDeviceError}</span>
+                    <Button
+                      variant="outline"
+                      size="sm"
+                      onClick={() => {
+                        dispatch(clearLocalDeviceError())
+                        dispatch(fetchLocalDeviceInfo())
+                      }}
+                    >
+                      {t('devices.list.actions.retry')}
+                    </Button>
+                  </AlertDescription>
+                </Alert>
+              </div>
+            ) : (
+              localDeviceLoading && <LocalPanelSkeleton />
+            ))}
+
+          {effectiveSelection.kind === 'peer' && selectedPeer && (
+            <PeerDetailPanel
+              key={selectedPeer.peerId}
+              deviceId={selectedPeer.peerId}
+              device={selectedPeer}
+              globalAutoSyncOff={globalAutoSyncOff}
+              globalFileSyncOff={globalFileSyncOff}
               lanOnlyActive={lanOnlyActive}
-              onSelect={handleSelectPeer}
-              onAdd={() => setAddP2PDialogOpen(true)}
+              onUnpair={handleUnpairRequest}
             />
           )}
-        </TabsContent>
 
-        <TabsContent value="mobile" className="mt-6">
-          {mobileDevicesError ? (
-            <Alert variant="destructive">
-              <AlertDescription className="flex items-center gap-3">
-                <span className="flex-1">{mobileDevicesError}</span>
-                <Button variant="ghost" size="icon-sm" onClick={mobileActions.reload}>
-                  <RefreshCw className="h-4 w-4" />
-                </Button>
-              </AlertDescription>
-            </Alert>
-          ) : (
-            <MobileGrid
-              mobiles={mobileDevices}
-              onSelect={mobileActions.openDetail}
-              onAdd={mobileActions.handleAddClick}
+          {effectiveSelection.kind === 'mobile' && selectedMobile && (
+            <MobileDevicePanel
+              key={selectedMobile.deviceId}
+              device={selectedMobile}
+              settings={mobileSettings}
+              initialCredential={
+                pendingCredential?.deviceId === selectedMobile.deviceId ? pendingCredential : null
+              }
+              onRevoke={mobileActions.requestRevoke}
+              onRotated={mobileActions.reload}
             />
           )}
-        </TabsContent>
-      </Tabs>
+        </ScrollArea>
+      </main>
 
-      {/* ── P2P dialogs ────────────────────────────────────────── */}
+      {/* ── flow dialogs ──────────────────────────────────────── */}
       <AddDeviceDialog open={addP2PDialogOpen} onOpenChange={setAddP2PDialogOpen} />
-      <DeviceSettingsDialog
-        open={peerDialogOpen}
-        onOpenChange={setPeerDialogOpen}
-        deviceId={selectedPeerId || ''}
-        device={selectedPeer}
-        globalAutoSyncOff={globalAutoSyncOff}
-        globalFileSyncOff={globalFileSyncOff}
-        lanOnlyActive={lanOnlyActive}
-        onUnpair={handleUnpairRequest}
-      />
       <UnpairAlertDialog
         open={unpairDialogOpen}
         onOpenChange={setUnpairDialogOpen}
         deviceName={unpairTargetDevice?.deviceName || t('devices.list.labels.unknownDevice')}
         onConfirm={handleUnpairConfirm}
       />
-
-      {/* ── Mobile dialogs / modals ────────────────────────────── */}
       <MobileSyncSettingsDialog
         open={settingsSheetOpen}
         onOpenChange={mobileActions.setSettingsSheetOpen}
@@ -570,21 +446,14 @@ const DeviceTabs: React.FC = () => {
       <AddMobileSyncDeviceDialog
         open={addDialogOpen}
         onOpenChange={mobileActions.setAddDialogOpen}
-        onSuccess={mobileActions.handleAddSuccess}
-      />
-      <MobileSyncDeviceDialog
-        open={mobileDetailDevice !== null}
-        onOpenChange={open => {
-          if (!open) mobileActions.closeDetail()
+        onSuccess={result => {
+          // Retire the credential modal: refresh the list, select the new
+          // device, and hand its one-time credentials to the panel's fresh
+          // state (pairing QR + credentials + install helper) inline.
+          mobileActions.reload()
+          setPendingCredential(result)
+          setSelection({ kind: 'mobile', id: result.deviceId })
         }}
-        device={mobileDetailDevice}
-        settings={mobileSettings}
-        onRevoke={mobileActions.requestRevoke}
-        onRotated={mobileActions.reload}
-      />
-      <MobileSyncCredentialModal
-        payload={credentialPayload}
-        onComplete={mobileActions.completeCredential}
       />
 
       <AlertDialog
@@ -618,240 +487,94 @@ const DeviceTabs: React.FC = () => {
           </AlertDialogFooter>
         </AlertDialogContent>
       </AlertDialog>
-    </>
-  )
-}
-
-const CountChip: React.FC<{ count: number; active: boolean }> = ({ count, active }) => (
-  <span
-    className={cn(
-      'inline-flex h-[18px] min-w-[18px] items-center justify-center rounded-full px-1.5 text-[10px] font-semibold tabular-nums transition-colors',
-      active ? 'bg-foreground text-background' : 'bg-muted text-muted-foreground'
-    )}
-  >
-    {count}
-  </span>
-)
-
-// ────────────────────────────────────────────────────────────────
-// Peer 网格（P2P）
-// ────────────────────────────────────────────────────────────────
-
-interface PeerGridProps {
-  peers: SpaceMember[]
-  lanOnlyActive: boolean
-  onSelect: (peerId: string) => void
-  onAdd: () => void
-}
-
-const PeerGrid: React.FC<PeerGridProps> = ({ peers, lanOnlyActive, onSelect, onAdd }) => {
-  const { t } = useTranslation()
-  return (
-    <div className="grid grid-cols-1 gap-3 sm:grid-cols-2 xl:grid-cols-3">
-      {peers.map(peer => (
-        <PeerCard key={peer.peerId} peer={peer} lanOnlyActive={lanOnlyActive} onSelect={onSelect} />
-      ))}
-      <AddCard
-        label={t('devices.addDevice.title')}
-        hint={t('devices.addDevice.cardHint', {
-          defaultValue: '生成邀请码 · 跨设备配对',
-        })}
-        onClick={onAdd}
-      />
     </div>
   )
 }
 
-interface PeerCardProps {
-  peer: SpaceMember
-  lanOnlyActive: boolean
-  onSelect: (peerId: string) => void
-}
-
-type ChannelTone = 'emerald' | 'amber' | 'muted'
-
-const CHANNEL_ICON: Record<
-  ReturnType<typeof deriveBadgeKind>,
-  { icon: typeof Wifi; tone: ChannelTone }
-> = {
-  lan: { icon: Cable, tone: 'emerald' },
-  relay: { icon: Wifi, tone: 'amber' },
-  offline: { icon: WifiOff, tone: 'muted' },
-  unknown: { icon: WifiOff, tone: 'muted' },
-  outOfLan: { icon: WifiOff, tone: 'amber' },
-}
-
-const PeerCard = React.memo(({ peer, lanOnlyActive, onSelect }: PeerCardProps) => {
-  const { t } = useTranslation()
-  const Icon = getDeviceIcon(peer.deviceName)
-  const kind = deriveBadgeKind(peer.channel ?? 'unknown', lanOnlyActive)
-  const channel = CHANNEL_ICON[kind]
-
-  return (
-    <button
-      type="button"
-      onClick={() => onSelect(peer.peerId)}
-      className="group relative flex w-full flex-col gap-4 overflow-hidden rounded-2xl border border-border/60 bg-card p-5 text-left transition-all hover:border-border hover:shadow-sm focus-visible:border-ring focus-visible:ring-3 focus-visible:ring-ring/50 focus-visible:outline-none"
-    >
-      <div className="flex items-start justify-between">
-        <div
-          className={cn(
-            'relative flex h-12 w-12 items-center justify-center rounded-xl',
-            peer.connected ? 'bg-success/10 text-success' : 'bg-muted text-muted-foreground'
-          )}
-        >
-          {/* eslint-disable-next-line react-hooks/static-components -- `getDeviceIcon` returns a stable lucide icon reference keyed on deviceName, not a freshly-created component */}
-          <Icon className="h-6 w-6" />
-          {peer.connected && (
-            <span className="absolute -bottom-0.5 -right-0.5 h-3 w-3 rounded-full bg-success ring-2 ring-card" />
-          )}
-        </div>
-        <ChevronRight className="h-4 w-4 text-muted-foreground/30 transition-all group-hover:translate-x-0.5 group-hover:text-muted-foreground" />
-      </div>
-
-      <div className="min-w-0">
-        <h4 className="truncate text-base font-semibold leading-tight text-foreground">
-          {peer.deviceName || t('devices.list.labels.unknownDevice')}
-        </h4>
-        <p className="mt-1 truncate text-xs text-muted-foreground">
-          <span className={peer.connected ? 'text-success' : 'text-muted-foreground'}>
-            {peer.connected
-              ? `● ${t('devices.list.status.online')}`
-              : `○ ${t('devices.list.status.offline')}`}
-          </span>
-          {peer.connectionAddress && (
-            <>
-              <span className="mx-1.5 text-muted-foreground/40">·</span>
-              <span className="font-mono">{peer.connectionAddress}</span>
-            </>
-          )}
-        </p>
-      </div>
-
-      <div className="pt-1">
-        <ChannelChip
-          icon={channel.icon}
-          label={t(`devices.list.channel.${kind}`)}
-          tone={channel.tone}
-        />
-      </div>
-    </button>
-  )
-})
-PeerCard.displayName = 'PeerCard'
-
-interface ChannelChipProps {
-  icon: typeof Wifi
-  label: string
-  tone: ChannelTone
-}
-
-const CHANNEL_TONE: Record<ChannelTone, string> = {
-  emerald: 'bg-success/10 text-success ring-success/20',
-  amber: 'bg-warning/10 text-warning ring-warning/20',
-  muted: 'bg-muted text-muted-foreground ring-border',
-}
-
-const ChannelChip: React.FC<ChannelChipProps> = ({ icon: Icon, label, tone }) => (
-  <span
-    className={cn(
-      'inline-flex items-center gap-1 rounded-md px-2 py-1 text-[11px] font-medium ring-1',
-      CHANNEL_TONE[tone]
-    )}
-  >
-    <Icon className="h-3 w-3" />
-    {label}
-  </span>
-)
+export default DevicesPage
 
 // ────────────────────────────────────────────────────────────────
-// Mobile 网格
+// List column pieces
 // ────────────────────────────────────────────────────────────────
 
-interface MobileGridProps {
-  mobiles: MobileDeviceView[]
-  onSelect: (device: MobileDeviceView) => void
-  onAdd: () => void
-}
-
-const MobileGrid: React.FC<MobileGridProps> = ({ mobiles, onSelect, onAdd }) => {
-  const { t } = useTranslation()
-  return (
-    <div className="grid grid-cols-1 gap-3 sm:grid-cols-2 xl:grid-cols-3">
-      {mobiles.map(mobile => (
-        <MobileCard key={mobile.deviceId} mobile={mobile} onSelect={() => onSelect(mobile)} />
-      ))}
-      <AddCard
-        label={t('devices.mobileSync.add.title')}
-        hint={t('devices.mobileSync.add.cardHint', {
-          defaultValue: '生成凭据 · 适用 iOS / Android',
-        })}
-        onClick={onAdd}
-      />
-    </div>
-  )
-}
-
-const MobileCard: React.FC<{
-  mobile: MobileDeviceView
-  onSelect: () => void
-}> = ({ mobile, onSelect }) => {
-  const { t } = useTranslation()
-  const lastSeen = formatLastSeen(mobile.lastSeenAtMs ?? null, t)
-
-  return (
-    <button
-      type="button"
-      onClick={onSelect}
-      className="group relative flex w-full flex-col gap-4 overflow-hidden rounded-2xl border border-border/60 bg-card p-5 text-left transition-all hover:border-border hover:shadow-sm focus-visible:border-ring focus-visible:ring-3 focus-visible:ring-ring/50 focus-visible:outline-none"
-    >
-      <div className="flex items-start justify-between">
-        <div className="flex h-12 w-12 items-center justify-center rounded-xl bg-info/10 text-info">
-          <Smartphone className="h-6 w-6" />
-        </div>
-        <ChevronRight className="h-4 w-4 text-muted-foreground/30 transition-all group-hover:translate-x-0.5 group-hover:text-muted-foreground" />
-      </div>
-
-      <div className="min-w-0">
-        <h4 className="truncate text-base font-semibold leading-tight text-foreground">
-          {mobile.label}
-        </h4>
-        <p className="mt-1 truncate font-mono text-xs text-muted-foreground">{mobile.username}</p>
-      </div>
-
-      <div className="pt-1">
-        <span className="text-[11px] text-muted-foreground">{lastSeen}</span>
-      </div>
-    </button>
-  )
-}
-
-// ────────────────────────────────────────────────────────────────
-// Add card (虚线占位)
-// ────────────────────────────────────────────────────────────────
-
-const AddCard: React.FC<{ label: string; hint: string; onClick: () => void }> = ({
+const SectionLabel: React.FC<{ label: string; trailing?: React.ReactNode }> = ({
   label,
-  hint,
-  onClick,
+  trailing,
+}) => (
+  <div className="flex items-center justify-between px-2.5 pb-1 pt-4">
+    <span className="text-[10px] font-semibold uppercase tracking-[0.12em] text-muted-foreground/80">
+      {label}
+    </span>
+    {trailing}
+  </div>
+)
+
+interface DeviceListItemProps {
+  name: string
+  tone: StatusDotTone
+  selected: boolean
+  dimmed?: boolean
+  trailing?: string
+  onSelect: () => void
+}
+
+const DeviceListItem: React.FC<DeviceListItemProps> = ({
+  name,
+  tone,
+  selected,
+  dimmed,
+  trailing,
+  onSelect,
 }) => (
   <button
     type="button"
-    onClick={onClick}
-    className="group flex flex-col items-center justify-center gap-3 rounded-2xl border-2 border-dashed border-border/60 bg-transparent p-5 text-center transition-all hover:border-primary/40 hover:bg-primary/5"
+    onClick={onSelect}
+    className={cn(
+      'flex w-full items-center gap-2.5 rounded-lg px-2.5 py-2 text-left text-sm transition-colors',
+      selected
+        ? 'bg-muted font-semibold text-foreground'
+        : 'font-medium text-foreground hover:bg-muted/60',
+      dimmed && !selected && 'opacity-60'
+    )}
   >
-    <div className="flex h-12 w-12 items-center justify-center rounded-xl bg-muted text-muted-foreground transition-colors group-hover:bg-primary/10 group-hover:text-primary">
-      <Plus className="h-6 w-6" />
-    </div>
-    <div>
-      <p className="text-sm font-medium text-foreground">{label}</p>
-      <p className="mt-1 text-[11px] text-muted-foreground">{hint}</p>
-    </div>
+    <StatusDot tone={tone} />
+    <span className="min-w-0 flex-1 truncate">{name}</span>
+    {trailing && (
+      <span className="shrink-0 text-[10px] font-normal text-muted-foreground">{trailing}</span>
+    )}
   </button>
 )
 
+const LocalPanelSkeleton: React.FC = () => (
+  <div className="mx-auto flex w-full max-w-2xl flex-col gap-6 px-8 py-8">
+    <div className="flex items-center gap-4">
+      <Skeleton className="size-12 rounded-xl" />
+      <div className="flex flex-1 flex-col gap-2">
+        <Skeleton className="h-5 w-44" />
+        <Skeleton className="h-3 w-28" />
+      </div>
+    </div>
+    <Skeleton className="h-16 w-full" />
+  </div>
+)
+
 // ────────────────────────────────────────────────────────────────
-// Mobile devices hook —— 集中存放原 MobileSyncDevicesPanel 的状态机
+// Presence helpers
+// ────────────────────────────────────────────────────────────────
+
+function peerDotTone(peer: SpaceMember, lanOnlyActive: boolean): StatusDotTone {
+  if (!peer.connected) return 'off'
+  const kind = deriveBadgeKind(peer.channel ?? 'unknown', lanOnlyActive)
+  return kind === 'relay' ? 'warning' : 'success'
+}
+
+function mobileDotTone(mobile: MobileDeviceView): StatusDotTone {
+  if (mobile.lastSeenAtMs == null) return 'off'
+  return Date.now() - mobile.lastSeenAtMs <= MOBILE_ACTIVE_WINDOW_MS ? 'info' : 'off'
+}
+
+// ────────────────────────────────────────────────────────────────
+// Mobile devices hook (state machine shared by list + panels + dialogs)
 // ────────────────────────────────────────────────────────────────
 
 interface UseMobileDevicesReturn {
@@ -861,27 +584,20 @@ interface UseMobileDevicesReturn {
   addDialogOpen: boolean
   settingsSheetOpen: boolean
   enableConfirmOpen: boolean
-  credentialPayload: RegisterMobileDeviceResult | null
   revokeTarget: MobileDeviceView | null
   revokeBusy: boolean
-  /** 当前在 MobileSyncDeviceDialog 里查看的设备;null = dialog 关闭。 */
-  detailDevice: MobileDeviceView | null
   actions: {
     reload: () => void
     handleAddClick: () => void
     handleEnableSuccess: () => void
-    handleAddSuccess: (result: RegisterMobileDeviceResult) => void
     handleRevokeConfirm: () => Promise<void>
     requestRevoke: (device: MobileDeviceView) => void
-    completeCredential: () => void
     clearRevokeTarget: () => void
     setAddDialogOpen: (open: boolean) => void
     setSettingsSheetOpen: (open: boolean) => void
     setEnableConfirmOpen: (open: boolean) => void
     setSettings: (settings: MobileSyncSettingsView | null) => void
     openSettings: () => void
-    openDetail: (device: MobileDeviceView) => void
-    closeDetail: () => void
   }
 }
 
@@ -895,18 +611,9 @@ const useMobileDevices = (): UseMobileDevicesReturn => {
   const [settingsSheetOpen, setSettingsSheetOpen] = useState(false)
   const [addDialogOpen, setAddDialogOpen] = useState(false)
   const [enableConfirmOpen, setEnableConfirmOpen] = useState(false)
-  const [credentialPayload, setCredentialPayload] = useState<RegisterMobileDeviceResult | null>(
-    null
-  )
 
   const [revokeTarget, setRevokeTarget] = useState<MobileDeviceView | null>(null)
   const [revokeBusy, setRevokeBusy] = useState(false)
-
-  // detail dialog 不存 deviceId 而存整个 view; 列表 reload 后通过 deviceId
-  // 重新 reconcile, 保证 dialog 内的字段始终是最新的服务端快照(关键是
-  // lastSeen / reportedOs 在用户开 dialog 期间可能被 ws 事件更新)。
-  // 改密 / 改密结果都在 dialog 内部组件状态里, 不再上提到 hook。
-  const [detailDeviceId, setDetailDeviceId] = useState<string | null>(null)
 
   const translate = useCallback((err: unknown): string => translateMobileSyncError(t, err), [t])
 
@@ -921,10 +628,10 @@ const useMobileDevices = (): UseMobileDevicesReturn => {
     }
   }, [translate])
 
-  // 首屏拉一次 settings：Add 按钮的 disable / 引导对话框分支都依赖
-  // settings.enabled / lanListenerError，settings 只在 Settings 对话框
-  // 打开时才会被回灌，初次进页面前必须主动取一次，否则 Add 永远走
-  // EnableConfirm 流程、bind 失败的硬阻断也失效。
+  // Preload settings on first render: the add flow's gating (enable
+  // confirm / bind-error hard block) reads settings.enabled and
+  // settings.lanListenerError, which are otherwise only refreshed when
+  // the settings dialog opens.
   useEffect(() => {
     void reload()
     getMobileSyncSettings()
@@ -935,14 +642,16 @@ const useMobileDevices = (): UseMobileDevicesReturn => {
   }, [reload])
 
   const handleAddClick = useCallback(() => {
-    // bind 失败仍是硬阻断 —— 没起监听就没法接 iPhone, 任何 add 都没意义。
+    // A failed LAN bind is a hard block: without a listener no mobile
+    // device can connect, so adding one is pointless.
     if (settings?.lanListenerError) {
       toast.error(
         t('devices.mobileSync.statusBar.bindFailed', { reason: settings.lanListenerError })
       )
       return
     }
-    // 首次入口：未开启 enabled / lan_listen → 弹引导对话框,确认后再 Add。
+    // First-run path: mobile sync or the LAN listener is off, walk the
+    // user through the enable confirm before the add dialog.
     if (!settings?.enabled || !settings?.lanListenEnabled) {
       setEnableConfirmOpen(true)
       return
@@ -952,21 +661,6 @@ const useMobileDevices = (): UseMobileDevicesReturn => {
 
   const handleEnableSuccess = useCallback(() => {
     setAddDialogOpen(true)
-  }, [])
-
-  const handleAddSuccess = useCallback(
-    (result: RegisterMobileDeviceResult) => {
-      setCredentialPayload(result)
-      void reload()
-    },
-    [reload]
-  )
-
-  // "撤销刚注册的设备"已下沉到设备卡片的 revoke 按钮(handleRevokeConfirm)。
-  // 凭据 modal 现在只承担凭据展示 + 配对引导, 关闭路径统一收敛到
-  // completeCredential — 用户后悔时去设备卡片删, 与"在 modal 里点 X"分流。
-  const completeCredential = useCallback(() => {
-    setCredentialPayload(null)
   }, [])
 
   const handleRevokeConfirm = useCallback(async () => {
@@ -985,11 +679,6 @@ const useMobileDevices = (): UseMobileDevicesReturn => {
     }
   }, [reload, revokeTarget, t, translate])
 
-  const detailDevice = useMemo(
-    () => (detailDeviceId ? (devices.find(d => d.deviceId === detailDeviceId) ?? null) : null),
-    [detailDeviceId, devices]
-  )
-
   return {
     devices,
     devicesError,
@@ -997,32 +686,26 @@ const useMobileDevices = (): UseMobileDevicesReturn => {
     addDialogOpen,
     settingsSheetOpen,
     enableConfirmOpen,
-    credentialPayload,
     revokeTarget,
     revokeBusy,
-    detailDevice,
     actions: {
       reload: () => void reload(),
       handleAddClick,
       handleEnableSuccess,
-      handleAddSuccess,
       handleRevokeConfirm,
       requestRevoke: setRevokeTarget,
-      completeCredential,
       clearRevokeTarget: () => setRevokeTarget(null),
       setAddDialogOpen,
       setSettingsSheetOpen,
       setEnableConfirmOpen,
       setSettings,
       openSettings: () => setSettingsSheetOpen(true),
-      openDetail: (device: MobileDeviceView) => setDetailDeviceId(device.deviceId),
-      closeDetail: () => setDetailDeviceId(null),
     },
   }
 }
 
 // ────────────────────────────────────────────────────────────────
-// 辅助函数
+// Helpers
 // ────────────────────────────────────────────────────────────────
 
 /**
@@ -1066,23 +749,4 @@ function translateMobileSyncError(t: ReturnType<typeof useTranslation>['t'], err
   }
   const message = err instanceof Error ? err.message : String(err)
   return t('devices.mobileSync.errors.unknown', { message })
-}
-
-function formatLastSeen(
-  lastSeenAtMs: number | null,
-  t: ReturnType<typeof useTranslation>['t']
-): string {
-  if (lastSeenAtMs == null)
-    return t('devices.mobileSync.list.lastSeen.never', { defaultValue: '从未活动' })
-  const diffMs = Date.now() - lastSeenAtMs
-  const diffMins = Math.round(diffMs / 60000)
-  let rel: string
-  if (diffMins < 1) rel = t('devices.mobileSync.list.lastSeen.justNow', { defaultValue: '刚刚' })
-  else if (diffMins < 60) rel = `${diffMins}m`
-  else if (diffMins < 1440) rel = `${Math.floor(diffMins / 60)}h`
-  else rel = `${Math.floor(diffMins / 1440)}d`
-  return t('devices.mobileSync.list.lastSeen.label', {
-    defaultValue: '最近活动 · {{rel}}',
-    rel,
-  })
 }
